% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_running_aux.R
\name{fire_interrupt}
\alias{fire_interrupt}
\title{Função que altera o status atual da tarefa para "interrupt".}
\usage{
fire_interrupt(status_file)
}
\arguments{
\item{status_file}{Um arquivo temporário onde está registrado o status da tarefa. Usualmente criado com o comando `status_file = tempfile()`.}
}
\value{
None
}
\description{
Usada como auxiliar para interromper a tarefa.
}
