% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_running_aux.R
\name{fire_ready}
\alias{fire_ready}
\title{Função que altera o status atual da tarefa para "Ready".}
\usage{
fire_ready(status_file)
}
\arguments{
\item{status_file}{Um arquivo temporário onde está registrado o status da tarefa. Usualmente criado com o comando `status_file = tempfile()`.}
}
\value{
None
}
\description{
Usada como auxiliar para avisar que a tarefa terminou e permitir que seja executada novamente.
}
