% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_running_aux.R
\name{fire_running}
\alias{fire_running}
\title{Função que altera o status atual da tarefa para "Running..." e opicionalmente informa a porcentagem da tarefa que foi concluída.}
\usage{
fire_running(perc_complete, status_file)
}
\arguments{
\item{perc_complete}{Um número entre 0 e 100 que representa a procentagem da tarefa que já foi rodada.}

\item{status_file}{Um arquivo temporário onde está registrado o status da tarefa. Usualmente criado com o comando `status_file = tempfile()`.}
}
\value{
None
}
\description{
Usada como auxiliar para avisar que a tarefa está rodando e opcionalmente a porcentagem completa.
}
