% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_running_aux.R
\name{get_status}
\alias{get_status}
\title{Função que pega o status atual da tarefa.}
\usage{
get_status(status_file)
}
\arguments{
\item{status_file}{Um arquivo temporário onde está registrado o status da tarefa. Usualmente criado com o comando `status_file = tempfile()`.}
}
\value{
O status atual assim como registrado no arquivo.
}
\description{
Função que pega o status atual da tarefa.
}
