% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_running_aux.R
\name{interrupted}
\alias{interrupted}
\title{Função que verifica se o status atual é igual a "interrupt".}
\usage{
interrupted(status_file)
}
\arguments{
\item{status_file}{Um arquivo temporário onde está registrado o status da tarefa. Usualmente criado com o comando `status_file = tempfile()`.}
}
\value{
None
}
\description{
Usada como auxiliar para interromper a tarefa que está rodando.
}
