% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_running_aux.R
\name{set_status}
\alias{set_status}
\title{Função que auxiliar usada para alterar o status atual da tarefa.}
\usage{
set_status(msg, status_file)
}
\arguments{
\item{msg}{String contendo o status da tarega.}

\item{status_file}{Um arquivo temporário onde está registrado o status da tarefa. Usualmente criado com o comando `status_file = tempfile()`.}
}
\value{
None
}
\description{
Função que auxiliar usada para alterar o status atual da tarefa.
}
